requirements.txt
PyYAML

pytest
pylint
autopep8
